improve language selection
Now you can define a keyword to specify the language. This prevents you have having to have a blank cell at the top in the language you want to define the language, e.g. if the first real cell is a shell command, but you want an ipython notebook, or if you have an all text notebook but want it to be an R or julia notebook, you can use the keyword.