add title cell for issue #39
I have wanted this for a while, and today I found a solution. I think it could
be more general maybe. This solution is hard coded to use the title, author and
date when found. I didn't see a simple solution to make it more flexible yet,
that would require some kind of template with optional things in it. Maybe in
the future...